Update project name
