+++ Makefile
@@ -1,2 +1,2 @@
 run:
-	pipenv run python -m src
+	pipenv run python -m goat_tweet_analyzer
