+++ Pipfile
@@ -10,6 +10,7 @@ ipdb = "*"
 [packages]
 loafer = "*"
 prettyconf = "*"
+requests = "*"
 
 [requires]
 python_version = "3.8"
