+++ goat_tweet_analyzer/translators.py
